+++ packages/instantsearch.js/src/components/SearchBox/SearchBox.tsx
@@ -195,7 +195,7 @@ class SearchBox extends Component<
             rootProps={{
               className: cssClasses.submit,
               type: 'submit',
-              title: 'Submit the search query.',
+              title: 'Submit the search query',
               hidden: !showSubmit,
             }}
             templates={templates}
@@ -208,7 +208,7 @@ class SearchBox extends Component<
             rootProps={{
               className: cssClasses.reset,
               type: 'reset',
-              title: 'Clear the search query.',
+              title: 'Clear the search query',
               hidden: !(
                 showReset &&
                 this.state.query.trim() &&