+++ packages/react-instantsearch/src/__tests__/common-widgets.test.tsx
@@ -22,6 +22,7 @@ import {
   HitsPerPage,
   ClearRefinements,
   CurrentRefinements,
+  ToggleRefinement,
 } from '..';
 
 import type { TestOptionsMap, TestSetupsMap } from '@instantsearch/tests';
@@ -266,6 +267,25 @@ const testSetups: TestSetupsMap<TestSuites> = {
       </form>
     );
   },
+  createToggleRefinementWidgetTests({ instantSearchOptions, widgetParams }) {
+    render(
+      <InstantSearch {...instantSearchOptions}>
+        <ToggleRefinement {...widgetParams} />
+        <GlobalErrorSwallower />
+      </InstantSearch>
+    );
+  },
+  createSearchBoxWidgetTests({
+    instantSearchOptions,
+    widgetParams: { autofocus, ...rest },
+  }) {
+    render(
+      <InstantSearch {...instantSearchOptions}>
+        <SearchBox {...rest} autoFocus={autofocus} />
+        <GlobalErrorSwallower />
+      </InstantSearch>
+    );
+  },
 };
 
 const testOptions: TestOptionsMap<TestSuites> = {
@@ -281,6 +301,8 @@ const testOptions: TestOptionsMap<TestSuites> = {
   createHitsPerPageWidgetTests: { act },
   createClearRefinementsWidgetTests: { act },
   createCurrentRefinementsWidgetTests: { act },
+  createToggleRefinementWidgetTests: { act },
+  createSearchBoxWidgetTests: { act },
 };
 
 /**