+++ packages/react-instantsearch/src/ui/__tests__/RefinementList.test.tsx
@@ -249,8 +249,10 @@ describe('RefinementList', () => {
                 action=""
                 class="ais-SearchBox-form"
                 novalidate=""
+                role="search"
               >
                 <input
+                  aria-label="Search"
                   autocapitalize="off"
                   autocomplete="off"
                   autocorrect="off"
@@ -302,6 +304,7 @@ describe('RefinementList', () => {
                 >
                   <svg
                     aria-hidden="true"
+                    aria-label="Results are loading"
                     class="ais-SearchBox-loadingIcon"
                     height="16"
                     stroke="#444"