+++ tests/common/widgets/index.ts
@@ -10,3 +10,5 @@ export * from './pagination';
 export * from './range-input';
 export * from './refinement-list';
 export * from './hits-per-page';
+export * from './toggle-refinement';
+export * from './search-box';