+++ packages/create-instantsearch-app/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.3.0](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.2.1...create-instantsearch-app@7.3.0) (2023-10-26)
+
+
+### Features
+
+* prompt users to enable support for events ([#5906](https://github.com/algolia/instantsearch/issues/5906)) ([675f1a0](https://github.com/algolia/instantsearch/commit/675f1a06d9ce9b8452d14344d6f3de30eb766e6c))
+
+
+
+
+
 ## [7.2.1](https://github.com/algolia/instantsearch/compare/create-instantsearch-app@7.2.0...create-instantsearch-app@7.2.1) (2023-10-24)
 
 ## Bug Fixes